[Caroussel][s]: moves component to another folder -- refs #30
